Rename to Doozer
Doozer is a community-owned fork of Henson. As part of this, all code
and documentation are being updated to reflect the new name. As part of
the documentation update, references to Sesame Street are being replaced
with references to Fraggle Rock.

As part of this fork, Doozer is being relicensed under the MIT license.
The LICENSE file is being updated to include a copy of the new license
in addition to the original license (along with a year correction in the
original license).

Note: pytest is being restricted to a previous version due to an error
that occurs using the latest version (3.3). This will be addressed in a
future change.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>