Update MANIFEST.in
Most of these changes are based on the output from check-manifest. The
`.gitkeep` file is being removed from `docs/_static` because the folder
contains the logo and will always exist (see below for more
information).

In addition to just cleaning up `MANIFEST.in`, this will now include the
tests and all files needed to generate the docs.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>