Better utilize the Read the Docs config file
I had envisioned this commit as being a much larger change, but it turns
out most of the settings used to generate Doozer's docs are the default.
The only setting support by the config file that doesn't use the default
(and wasn't already in the file) is `requirements_file`. While this is
currently set in the Read the Docs admin, having it set here will help
this file better serve as an example for other Doozer-related projects.

Signed-off-by: Andy Dirnberger <andy@dirnberger.me>