Ignore D401 where it doesn't make sense
The final step to allowing the pydocstyle upgrade in 175a404 is to
address the lack of imperative docstrings in a few places. The use of
imperatives in these places don't make sense as two of them are
properties and the other is part of a protocol and meant to serve to
describe the protocol, not the method's implementation.