Only run Python version-specific unit tests in CI
To simplify running and maintaining the tests, [tox] will only run the
unit tests (now under the `unit` environment) using the version of
Python under which tox is installed. The tests will be run for all
supported versions of Python, though, as part of CI. To do this, the
version-specific [GitHub Actions] jobs are being reduced to a single job
using the matrix strategy.

[github actions]: https://docs.github.com/en/free-pro-team@latest/actions
[tox]: https://tox.readthedocs.io