Test more things with more Python versions in CI
Historically only the unit tests have been run with all supported
versions of Python. The jobs to lint the code and to test that the
documentation builds were intentionally run using the minimum supported
version to ensure that new features of the language didn't creep into
the code. This left a hole in the coverage, though, as not all things
were tested using all supported versions of Python. By utilizing GitHub
Action's matrix strategy (just as was done in 6088490), we can address
this.

Note that the `manifest` job is being left as-is because it lints the
`MANIFEST.in` file, not the code itself (or anything about Python
directly).