Let environment control Python version in tox
Controlling the Python version is set in both the CI configuration and
the [tox] configuration. The former should be the source of truth as it
is the (testing) gatekeeper to getting a change merged. This can be
accomplished by removing the `basepython` value from tox environments
that set it.

For the environments that don't run under CI,  `format` and `release`,
`basepython` could have been left, but it's less important for the
latter and the former will soon have a [better way] of being handled.

[better way]: https://pre-commit.com
[tox]: https://tox.readthedocs.io