Postpone evaluation of annotations
[PEP 563] introduced a change to how annotations are evaluated. By
evaluating them later, forward references become easier to work with and
the cost of evaluating the annotations is no longer incurred at import
time. Because this behavior won't be the default until Python 3.10,
adding a `__future__` import to each Python module will enable it until
3.10 is the lowest supported version. [isort] has an option to
automatically add specified imports to any file it processes. This is
being turned on to automatically do this.

[isort]: https://pycqa.github.io/isort/
[pep 563]: https://www.python.org/dev/peps/pep-0563/